pyproject.toml
[tool.poetry]
name = "junior-ai-engineer-test"
version = "0.1.0"
description = ""
authors = ["Hsu-Pei-Chun <philosophysis@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
openai = "^1.61.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
